last warning on linux

+++ modules/surface_matching/src/t_hash_int.cpp
@@ -320,7 +320,7 @@ void hashtablePrint(hashtable_int *hashtbl)
         for (node=hashtbl->nodes[n]; node; node=next)
         {
             next = node->next;
-            printf("Key : %d, Data : %d\n", node->key, (size_t)node->data);
+            printf("Key : %d, Data : %d\n", node->key, (int)node->data);
         }
     }
 }
